requirements.txt
Flask==3.0.3
gunicorn==22.0.0
tensorflow==2.16.2
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
openai==0.28.1
python-dotenv==1.0.0